add thumbnail servlet, sample data

+++ src/main/java/org/moviereview/model/Actor.java
@@ -14,17 +14,17 @@ public class Actor implements Serializable
 {
    private static final long serialVersionUID = -5577759231154572526L;
    
-   @Id private Integer id;
+   @Id private Long id;
    private String firstName;
    private String lastName;
    private Date dob;
 
-   public Integer getId()
+   public Long getId()
    {
       return id;
    }
    
-   public void setId(Integer id)
+   public void setId(Long id)
    {
       this.id = id;
    }
